requirements.txt
asgiref==3.3.4
django==3.2.3
django-heroku==0.3.1
django-configurations==2.2
dj-database-url==0.5.0
django-debug-toolbar==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
django-crispy-forms==1.12.0
psycopg2==2.8.6
sqlparse==0.4.1
six==1.15.0
Pillow==8.1.2
pytx
